requirements.txt
# python3.10
streamlit # Version: 1.38.0
python_dotenv # Version: 1.0.1
google-generativeai # Version: 0.8.1
groq # Version: 0.11.0
PyPDF2 # Version: 3.0.1
fpdf # Version: 1.7.2
python-docx # Version: 1.1.2
streamlit-js-eval # Version: 0.1.7
